requirements.txt
rich>=13.0.0
requests>=2.31.0
dnspython>=2.4.0
python-whois>=0.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
PyYAML>=6.0
aiohttp>=3.9.0
aiofiles>=23.0.0
colorama>=0.4.6
jinja2>=3.1.0
tldextract>=5.1.0
shodan>=1.31.0
weasyprint>=60.0
